[opentracing] NewTracer in the tracer.go
